# openapi/openapi-sdtooling-tier1-v1.yaml
info:
  title: SD Tooling API
  description: |-
    The APIs described in the following document enable the management
    of the lifecycle for creating a Resource Description on the provider side, including
    its publication on the federated catalogue. The available APIs allow to define
    access and usage policies, specify the Resource Address on the provider side,
    and manage the validation and publication process on the federated catalogue.

    ## Business Context

    The API facilitates the creation and management of Resource Descriptions in a federated data space. It enables providers to publish metadata about their data assets on a federated catalogue, define access and usage policies, and manage the validation and publication workflow. This supports discoverability and interoperability across the data ecosystem.


    ### Key Actors

    - **Provider**: Creates and publishes Resource Descriptions, defines policies, and manages validation.

    - **Governance Authority**: Hosts the Federated Catalogue

    ### Typical Flow

    1. Create a Resource Description.

    2. Define access and usage policies.

    3. Specify the Resource Address.

    4. Submit the Resource Description for validation

    5. Publish the Resource Description on the federated catalogue
  contact:
    name: Simpl Programme
    url: https://simpl-programme.ec.europa.eu/
    email: cnect-simpl@ec.europa.eu
  license:
    name: European Union Public License (EUPL) 1.2
...
  version: "1.0"
servers:
...
security:
- bearerAuth: []
openapi: 3.0.1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: IAA cloud gateway JWT token
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InputObject:
      type: object
      properties:
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
...
    InputValidationIssue:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        in:
          type: string
          enum:
          - body
          - path
          - query
          - header
        name:
          type: string
        value:
          type: object
        inputs:
          type: array
          items:
            $ref: '#/components/schemas/InputObject'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    InternalServerErrorProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
    NotFoundProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
    BadRequestProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        href:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        issues:
          type: array
          items:
            $ref: '#/components/schemas/InputValidationIssue'
        additionalProperties:
          type: object
          additionalProperties:
            type: object
...
paths:
...
  /resourceDescriptions/{resourceDescriptionId}:
    get:
      tags:
      - Resource Description
      summary: Return details of given resource description stored in federated catalogue
      operationId: getResourceDescription
      parameters:
      - name: resourceDescriptionId
        in: path
        description: Resource description identifier (credentialSubject.@id)
        required: true
        schema:
          type: string
          example: did:web:registry.gaia-x.eu:DataOffering:123
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
              examples:
                returnValidResponse:
                  description: returnValidResponse
                  value:
                    claimsGraphUri:
                    - did:web:registry.gaia-x.eu:InfrastructureOffering:a94ef06c-a12d-4308-89aa-fffcd8a61126
                    offeringType: infrastructure
                    name: text title for infrastructure resource.
                    inLanguage: da
                    description: text description
                    serviceAccessPoint: http://serviceaccesspoint.com
  /resourceDescriptions/{resourceDescriptionId}/revoke:
    post:
      tags:
      - Resource Description
      summary: Revoke an already published self-description from the federated catalogue
      description: |
        Revokes a previously published resource description.
        After revocation, the self-description shall no longer be visible or discoverable in the federated catalogue.
      operationId: revokeResourceDescription
      parameters:
      - name: resourceDescriptionId
        in: path
        description: Resource description identifier (credentialSubject.@id)
        required: true
        schema:
          type: string
          example: did:web:registry.gaia-x.eu:DataOffering:123
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                BadRequestExample:
                  description: BadRequestExample
                  value:
                    type: urn:problem-type:simpl:badRequest
                    title: BadRequest
                    status: 400
                    instance: /resourceDescriptions/{resourceDescriptionId}/revoke
                    detail: Invalid resourceDescriptionId format
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                InternalServerErrorExample:
                  description: InternalServerErrorExample
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    instance: /resourceDescriptions/{resourceDescriptionId}/revoke
                    detail: Unexpected internal error
        "200":
          description: Successfully revoked self-description
          content:
            application/json:
              schema:
                type: object
              examples:
                RevokeSuccess:
                  description: RevokeSuccess
                  value:
                    id: did:web:registry.gaia-x.eu:DataOffering:3mwO1...
                    status: revoked
                    statusDatetime: 2025-03-20T10:31:00Z
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/NotFoundProblem'
              examples:
                NotFoundExample:
                  description: NotFoundExample
                  value:
                    type: resource-not-found
                    title: RESOURCE_DESCRIPTION_NOT_FOUND
                    status: 404
                    instance: /resourceDescriptions/{resourceDescriptionId}/revoke
                    detail: Resource description not found or already revoked
  /resourceDescriptions:
    get:
      tags:
      - Resource Description
      summary: Return all resource descriptions with the participant id as defined in the Client access token.
      operationId: getAllResourceDescriptions
      parameters:
      - name: orderBy
        in: query
        description: Field to order the results by
        required: false
        schema:
          type: string
          description: Field to order the results by
          example: publicationDate
          default: publicationDate
      responses:
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
              examples:
                Bad Request Error Example:
                  description: Bad Request Error Example
                  value:
                    type: urn:problem-type:simpl:validationError
                    title: Invalid payload
                    status: 400
                    detail: missing required arguments
            '*/*':
              schema:
                $ref: '#/components/schemas/BadRequestProblem'
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorProblem'
              examples:
                Internal Server Error Example:
                  description: Internal Server Error Example
                  value:
                    type: urn:problem-type:simpl:internalServerError
                    title: Internal Server Error
                    status: 500
                    detail: Unexpected internal error
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
              examples:
                returnValidResponse:
                  description: returnValidResponse
                  value:
                    totalCount: 1
                    items:
                    - "n":
                        claimsGraphUri:
                        - did:web:registry.gaia-x.eu:InfrastructureOffering:a94ef06c-a12d-4308-89aa-fffcd8a61126
                        offeringType: infrastructure
                        name: text title for infrastructure resource.
                        inLanguage: da
                        description: text description
                        serviceAccessPoint: http://serviceaccesspoint.com
...